Add admin users and management commands

+++ discord_key_bot/command/admin.py
@@ -1,28 +1,122 @@
 import datetime
 import inspect
 import logging
-from typing import Optional, Sequence
+import re
+from typing import Sequence, Optional
 
+import discord
+from discord import User
 from discord.ext import commands
 from discord.ext.commands import Bot
 from sqlalchemy.orm import sessionmaker, Session
 
-from discord_key_bot.common import util
+from discord_key_bot.command.util import is_admin
 from discord_key_bot.common.colours import Colours
-from discord_key_bot.common.util import get_search_name
+from discord_key_bot.common.util import get_search_name, embed, send_message
 from discord_key_bot.db import search
-from discord_key_bot.db.models import Game
+from discord_key_bot.db.models import Game, Member
 from discord_key_bot.platform import Platform, get_platform
 
 
 class AdminCommands(commands.Cog, name='Admin Commands', command_attrs=dict(hidden=True)):
-    def __init__(self, bot: Bot, db_session_maker: sessionmaker):
+    def __init__(self, bot: Bot, db_session_maker: sessionmaker, admin_role_id: int = 0):
         self.bot: Bot = bot
         self.db_sessionmaker: sessionmaker = db_session_maker
         self.logger = logging.getLogger(__name__)
+        self.admin_role_id = admin_role_id
+
+        self._member_patt = re.compile(r"<@(\d+)>")
+
+    @commands.is_owner()
+    @commands.command()
+    async def addadmin(
+        self,
+        ctx: commands.Context,
+        *,
+        member: str = commands.Parameter(
+            name="member",
+            displayed_name="Member",
+            description="Member to add as admin",
+            kind=inspect.Parameter.POSITIONAL_ONLY
+        ),
+    ):
+        """Add a member as an admin"""
+
+        self.logger.info(f"received addadmin request from user {ctx.author.display_name}")
+
+        user: User = await self._get_user(ctx, member)
+        if not user:
+            return
+
+        session: Session = self.db_sessionmaker()
+
+        member = Member.get(session, user.id, user.name)
+        member.is_admin = True
+
+        session.flush()
+        session.commit()
+
+        await ctx.author.send(embed=embed(f"Successfully added {user.name} as admin", colour=Colours.GREEN))
+
+    @commands.is_owner()
+    @commands.command()
+    async def rmadmin(
+        self,
+        ctx: commands.Context,
+        *,
+        member: str = commands.Parameter(
+            name="member",
+            displayed_name="Member",
+            description="Member to remove as admin",
+            kind=inspect.Parameter.POSITIONAL_ONLY
+        ),
+    ):
+        """Remove a member as an admin"""
+
+        self.logger.info(f"received rmadmin request from user {ctx.author.display_name}")
+
+        user: User = await self._get_user(ctx, member)
+        if not user:
+            return
+
+        session: Session = self.db_sessionmaker()
+
+        member = Member.get(session, user.id, user.name)
+        member.is_admin = False
+
+        session.flush()
+        session.commit()
+
+        await ctx.author.send(embed=embed(f"Successfully removed {user.name} as admin", colour=Colours.GREEN))
 
     @commands.command()
     @commands.is_owner()
+    async def lsadmin(self, ctx: commands.Context):
+        """List admin users"""
+
+        self.logger.info(f"received lsadmin request from user {ctx.author.display_name}")
+
+        session: Session = self.db_sessionmaker()
+
+        admin_users: Sequence[Member] = search.get_admin_members(session)
+
+        if not admin_users:
+            await ctx.author.send(embed=embed("No admin users found", colour=Colours.RED))
+            return
+
+        msg = embed(
+            title="Admin Users",
+            text="",
+            colour=Colours.GREEN
+        )
+
+        for admin in admin_users:
+            msg.add_field(name=admin.name,
+                          value=f"**id:** {admin.id}")
+
+        await ctx.author.send(embed=msg)
+
+    @commands.command()
     async def gameid(
         self,
         ctx: commands.Context,
@@ -40,15 +134,19 @@ class AdminCommands(commands.Cog, name='Admin Commands', command_attrs=dict(hidd
 
         session: Session = self.db_sessionmaker()
 
+        if not is_admin(session, ctx):
+            self.logger.info(f"{ctx.author.display_name} is not an authorized admin")
+            return
+
         games: Sequence[Game] = search.get_admin_games(
             session=session, game_name=game_name
         )
 
         if not games:
-            await ctx.author.send(embed=util.embed("Game not found", colour=Colours.RED))
+            await ctx.author.send(embed=embed("Game not found", colour=Colours.RED))
             return
 
-        msg = util.embed(
+        msg = embed(
             title="Game Info",
             text="",
             colour=Colours.GREEN
@@ -86,10 +184,15 @@ class AdminCommands(commands.Cog, name='Admin Commands', command_attrs=dict(hidd
 
         session: Session = self.db_sessionmaker()
 
+        if not is_admin(session, ctx):
+            self.logger.info(f"{ctx.author.display_name} is not an authorized admin")
+            return
+
+
         game = session.get(Game, game_id)
 
         if not game:
-            await ctx.author.send(embed=util.embed("Game not found", colour=Colours.RED))
+            await ctx.author.send(embed=embed("Game not found", colour=Colours.RED))
             return
 
         if game.name == get_search_name(new_name):
@@ -99,7 +202,7 @@ class AdminCommands(commands.Cog, name='Admin Commands', command_attrs=dict(hidd
 
             text: str = f"Renamed display name of existing game from '{game.pretty_name}' to '{new_name}'"
             self.logger.debug(text)
-            await ctx.author.send(embed=util.embed(title="Renamed game", text=text))
+            await ctx.author.send(embed=embed(title="Renamed game", text=text))
             return
 
         existing_game = search.get_game(session=session, game_name=new_name)
@@ -109,11 +212,11 @@ class AdminCommands(commands.Cog, name='Admin Commands', command_attrs=dict(hidd
             game.name = get_search_name(new_name)
 
             self.logger.debug(text)
-            await ctx.author.send(embed=util.embed(title="Renamed game", text=text))
+            await ctx.author.send(embed=embed(title="Renamed game", text=text))
         else:
             text: str = f"Moving keys from game ID {game.id} to game ID {existing_game.id}"
             self.logger.debug(text)
-            await ctx.author.send(embed=util.embed(title="Renaming game", text=text))
+            await ctx.author.send(embed=embed(title="Renaming game", text=text))
 
             for key in game.keys:
                 key.game_id = existing_game.id
@@ -125,7 +228,6 @@ class AdminCommands(commands.Cog, name='Admin Commands', command_attrs=dict(hidd
         session.commit()
 
     @commands.command()
-    @commands.is_owner()
     async def bulk_expire(
         self,
         ctx: commands.Context,
@@ -155,32 +257,36 @@ class AdminCommands(commands.Cog, name='Admin Commands', command_attrs=dict(hidd
 
         session: Session = self.db_sessionmaker()
 
+        if not is_admin(session, ctx):
+            self.logger.info(f"{ctx.author.display_name} is not an authorized admin")
+            return
+
         try:
             platform: Platform = get_platform(platform_name)
         except ValueError:
             await ctx.author.send(
-                embed=util.embed(f'"{platform_name}" is not valid platform', Colours.RED),
+                embed=embed(f'"{platform_name}" is not valid platform', Colours.RED),
             )
             return
 
         game = session.get(Game, game_id)
         if not game:
-            await ctx.author.send(embed=util.embed("Game not found", colour=Colours.RED))
+            await ctx.author.send(embed=embed("Game not found", colour=Colours.RED))
             return
 
         try:
             expiration_date = datetime.datetime.strptime(expiration, "%b %d %Y")
         except ValueError:
-            await util.send_message(
+            await send_message(
                 ctx=ctx,
-                msg=util.embed(f"Failed to parse expiration date.", Colours.RED),
+                msg=embed(f"Failed to parse expiration date.", Colours.RED),
             )
             return
 
         if expiration_date.date() <= datetime.datetime.now(datetime.UTC).date():
-            await util.send_message(
+            await send_message(
                 ctx=ctx,
-                msg=util.embed(f"Expiration date is in the past.", Colours.RED),
+                msg=embed(f"Expiration date is in the past.", Colours.RED),
             )
             return
 
@@ -191,19 +297,22 @@ class AdminCommands(commands.Cog, name='Admin Commands', command_attrs=dict(hidd
         session.flush()
         session.commit()
 
-        await util.send_message(
+        await send_message(
             ctx=ctx,
-            msg=util.embed(f"Set bulk expiration dates", Colours.RED),
+            msg=embed(f"Set bulk expiration dates", Colours.RED),
         )
 
     @commands.command()
-    @commands.is_owner()
     async def purge(self, ctx: commands.Context):
         """Purge expired keys and orphaned games"""
 
         self.logger.info(f"purge request from user {ctx.author.display_name}")
 
         session: Session = self.db_sessionmaker()
+
+        if not is_admin(session, ctx):
+            self.logger.info(f"{ctx.author.display_name} is not an authorized admin")
+            return
 
         game_count: int
         key_count: int
@@ -212,13 +321,12 @@ class AdminCommands(commands.Cog, name='Admin Commands', command_attrs=dict(hidd
         session.commit()
 
         await ctx.author.send(
-            embed=util.embed(
+            embed=embed(
                 title="Deleting Expired Keys",
                 text=f"{game_count} games, {key_count} keys deleted", colour=Colours.GREEN)
             )
 
     @commands.command()
-    @commands.is_owner()
     async def delete(
         self,
         ctx: commands.Context,
@@ -235,9 +343,13 @@ class AdminCommands(commands.Cog, name='Admin Commands', command_attrs=dict(hidd
 
         session: Session = self.db_sessionmaker()
 
+        if not is_admin(session, ctx):
+            self.logger.info(f"{ctx.author.display_name} is not an authorized admin")
+            return
+
         game = session.get(Game, game_id)
         if not game:
-            await ctx.author.send(embed=util.embed("Game not found", colour=Colours.RED))
+            await ctx.author.send(embed=embed("Game not found", colour=Colours.RED))
             return
 
         session.delete(game)
@@ -245,7 +357,27 @@ class AdminCommands(commands.Cog, name='Admin Commands', command_attrs=dict(hidd
         session.commit()
 
         await ctx.author.send(
-            embed=util.embed(
+            embed=embed(
                 title="Deleting Expired Keys",
                 text=f"game_id {game_id} deleted", colour=Colours.GREEN)
             )
+
+    async def _get_user(self, ctx: commands.Context, user_str: str) -> Optional[discord.User]:
+        match: re.Match = self._member_patt.match(user_str)
+        if not match:
+            await ctx.author.send(embed=embed("Invalid member", colour=Colours.RED))
+            return None
+
+        user_id: int
+        try:
+            user_id = int(match.group(1))
+        except ValueError:
+            await ctx.author.send(embed=embed("Invalid member ID", colour=Colours.RED))
+            return None
+
+        user: User = await self.bot.fetch_user(user_id)
+        if not user:
+            await ctx.author.send(embed=embed("No user found with the provided ID", colour=Colours.RED))
+            return None
+
+        return user
